certsum: Refactor into separate helper funcs
Break out core functionality into separate functions
dedicated to execution as goroutines:

- collectors (port scan results, cert check results)
- scanners (port scanner, cert scanner)

Still more work to do. This is mostly an attempt to break out
the code for easier review.

refs GH-135
